Refactor package structure (#256)
* Refactor package structure

- Refactor package so that measures can be imported/required
  independantly
- Move definitions index.ts file to all.ts
- Remove measure export from all.ts
- Fix import statements in tests
- Update tsconfig configs so that types are included in their own folder
  inside
- Update package.json to better support esm and typescript package
  requirements

* Fix linting errors

* Fix typo in test

* Add .js to all import statements

- Add .js to all import statements
- Update eslint rule so that it doesn't check imports. Typescript
  handles this already.

* Update Jest Config to use ESM modules

* Remove merge conflict lines (oof)
