+++ .npmignore
@@ -1,4 +1,3 @@
-**/__tests__
 coverage
 .eslintrc.js
 .github/**/*
