+++ src/definitions/acceleration.ts
@@ -1,4 +1,4 @@
-import { Measure, Unit } from './../index';
+import { Measure, Unit } from './../index.js';
 export type AccelerationUnits = AccelerationMetricUnits;
 export type AccelerationSystems = 'metric';
