+++ src/definitions/area.ts
@@ -1,4 +1,4 @@
-import { Measure, Unit } from './../index';
+import { Measure, Unit } from './../index.js';
 export type AreaUnits = AreaMetricUnits | AreaImperialUnits;
 export type AreaSystems = 'metric' | 'imperial';
