+++ src/definitions/charge.ts
@@ -1,4 +1,4 @@
-import { Measure, Unit } from './../index';
+import { Measure, Unit } from './../index.js';
 export type ChargeUnits = ChargeSIUnits;
 export type ChargeSystems = 'SI';
