+++ src/definitions/digital.ts
@@ -1,4 +1,4 @@
-import { Measure, Unit } from './../index';
+import { Measure, Unit } from './../index.js';
 export type DigitalUnits = DigitalBitsUnits | DigitalBytesUnits;
 export type DigitalSystems = 'bits' | 'bytes';
