+++ src/definitions/energy.ts
@@ -1,4 +1,4 @@
-import { Measure, Unit } from './../index';
+import { Measure, Unit } from './../index.js';
 export type EnergyUnits = EnergySIUnits;
 export type EnergySystems = 'SI';
