+++ src/definitions/frequency.ts
@@ -1,4 +1,4 @@
-import { Measure, Unit } from './../index';
+import { Measure, Unit } from './../index.js';
 export type FrequencyUnits = FrequencySIUnits;
 export type FrequencySystems = 'SI';
