+++ src/definitions/length.ts
@@ -1,4 +1,4 @@
-import { Measure, Unit } from './../index';
+import { Measure, Unit } from './../index.js';
 export type LengthUnits = LengthMetricUnits | LengthImperialUnits;
 export type LengthSystems = 'metric' | 'imperial';
