+++ src/definitions/pace.ts
@@ -1,4 +1,4 @@
-import { Measure, Unit } from './../index';
+import { Measure, Unit } from './../index.js';
 export type PaceUnits = PaceMetricUnits | PaceImperialUnits;
 export type PaceSystems = 'metric' | 'imperial';
