+++ src/definitions/reactivePower.ts
@@ -1,4 +1,4 @@
-import { Measure, Unit } from './../index';
+import { Measure, Unit } from './../index.js';
 export type ReactivePowerUnits = ReactivePowerSIUnits;
 export type ReactivePowerSystems = 'SI';
