+++ src/definitions/time.ts
@@ -1,4 +1,4 @@
-import { Measure, Unit } from './../index';
+import { Measure, Unit } from './../index.js';
 export type TimeUnits = TimeSIUnits;
 export type TimeSystems = 'SI';
