+++ tsconfig.json
@@ -5,7 +5,7 @@
     "allowJs": false,
     "checkJs": false,
     "noImplicitAny": true,
-    "declaration": true,
+    "declaration": false,
     "outDir": "./lib/esm",
     "strict": true,
     "esModuleInterop": true,
